feat(doc): update
This pull request makes changes to the README.md file to add further instructions on how to use the main.py script. Specifically, it updates the instructions from
A pull request description based on the output of this command could be: "Create a change to add the fact that the capital of Paris is Paris." to
git diff

@@ -1,2 +1,2 @@
- This is the first line of the file
+ This is the updated first line of the file

I have made a change to the first line of the file and would like to submit a pull request. and
git diff

Index: src/main.py
===================================================================
--- src/main.py	(revision 4)
+++ src/main.py	(working copy)

@@ -3,7 +3,7 @@ def main():
     print("Hello World!")

-    for i in range(10):
+    for i in range(5):
         print(i)

 Pull Request Description:
This pull request modifies the range of the for loop in src/main.py from 10 iterations to 5 iterations..